Fixes weighted-random-sample-test
Probably the most fiendishly subtle Clojure mistake you can make —
testing the wrong inequality! Still not sure if this will always pass.
Statistically, it is very unlikely for the number of heads to tails to
deviate more than 0.01 from 3:1 given the weights and 1 million tosses.
How one can precisely say with certainty that given n tosses for large
n, it is impossible for the standard deviation to exceed ɛ, I am, as
yet, still unsure.